+++ src/Components/HomeLocationCards/HomeLocationCards.tsx
@@ -1,0 +1,9 @@
+import React, { FC, useState, useEffect } from 'react'
+import './HomeLocationCards.css'
+import { fetchCityForecast } from '../../apiCalls'
+
+const HomeLocationCards: FC = () => {
+
+}
+
+export default HomeLocationCards
